rename controllers and update package structure for modularization
